chore: move all rules into legacy rules
