+++ src/legacy-rules/case-abbr.ts
