+++ src/legacy-rules/case-backslash.ts
