+++ src/legacy-rules/case-ellipsis.ts
