+++ src/legacy-rules/case-html-entity.ts
