+++ src/legacy-rules/case-linebreak.ts
