+++ src/legacy-rules/case-math-exp.ts
