+++ src/legacy-rules/case-traditional.ts
