+++ src/legacy-rules/mark-hyper.ts
