+++ src/legacy-rules/space-brackets.ts
