+++ src/legacy-rules/space-full-width-content.ts
