+++ src/legacy-rules/util.ts
