+++ src/run.ts
@@ -14,25 +14,25 @@ import hexo from './hypers/hexo'
 import vuepress from './hypers/vuepress'
 import md from './hypers/md'
 
-import markRaw from './rules/mark-raw'
-import markHyper from './rules/mark-hyper'
-import markHtml from './rules/mark-html'
-import unifyPunctuation from './rules/unify-punctuation'
-import caseAbbr from './rules/case-abbr'
-import spaceFullWidthContent from './rules/space-full-width-content'
-import spacePunctuation from './rules/space-punctuation'
-import caseMathExp from './rules/case-math-exp'
-import caseBackslash from './rules/case-backslash'
-import spaceBrackets from './rules/space-brackets'
-import spaceQuotes from './rules/space-quotes'
-import caseTraditional from './rules/case-traditional'
-import caseDatetime from './rules/case-datetime'
-import caseDatetimeZh from './rules/case-datetime-zh'
-import caseEllipsis from './rules/case-ellipsis'
-import caseHtmlEntity from './rules/case-html-entity'
-import caseRaw from './rules/case-raw'
-import caseLinebreak from './rules/case-linebreak'
-// import markRawReview from './rules/mark-raw-review'
+import markRaw from './legacy-rules/mark-raw'
+import markHyper from './legacy-rules/mark-hyper'
+import markHtml from './legacy-rules/mark-html'
+import unifyPunctuation from './legacy-rules/unify-punctuation'
+import caseAbbr from './legacy-rules/case-abbr'
+import spaceFullWidthContent from './legacy-rules/space-full-width-content'
+import spacePunctuation from './legacy-rules/space-punctuation'
+import caseMathExp from './legacy-rules/case-math-exp'
+import caseBackslash from './legacy-rules/case-backslash'
+import spaceBrackets from './legacy-rules/space-brackets'
+import spaceQuotes from './legacy-rules/space-quotes'
+import caseTraditional from './legacy-rules/case-traditional'
+import caseDatetime from './legacy-rules/case-datetime'
+import caseDatetimeZh from './legacy-rules/case-datetime-zh'
+import caseEllipsis from './legacy-rules/case-ellipsis'
+import caseHtmlEntity from './legacy-rules/case-html-entity'
+import caseRaw from './legacy-rules/case-raw'
+import caseLinebreak from './legacy-rules/case-linebreak'
+// import markRawReview from './legacy-rules/mark-raw-review'
 
 const hyperParseInfo = [
   { name: 'ignore', value: ignore },
